requirements.txt
python-telegram-bot==21.11.1
asyncpg>=0.30.0
gspread==6.1.4
google-auth==2.34.0
python-dotenv==1.0.1
